# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1/menus":{"get":{"tags":["menus"],"summary":"Get Menus List","description":"Получить список всех меню","operationId":"get_menus_list_api_v1_menus_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MenuOut"},"type":"array","title":"Response Get Menus List Api V1 Menus Get"}}}}}},"post":{"tags":["menus"],"summary":"Create Menu","description":"Создать меню","operationId":"create_menu_api_v1_menus_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuIn"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/menus/{menu_id}":{"get":{"tags":["menus"],"summary":"Get Menu","description":"Получить меню по id","operationId":"get_menu_api_v1_menus__menu_id__get","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","title":"Menu Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuOut"}}}},"404":{"description":"menu not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["menus"],"summary":"Update Menu","description":"Обновить меню","operationId":"update_menu_api_v1_menus__menu_id__patch","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","title":"Menu Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["menus"],"summary":"Delete Menu","description":"Удалить меню","operationId":"delete_menu_api_v1_menus__menu_id__delete","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","title":"Menu Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/menus/{menu_id}/submenus":{"get":{"tags":["submenus"],"summary":"Get Submenus","description":"Получить список всех подменю","operationId":"get_submenus_api_v1_menus__menu_id__submenus_get","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","title":"Menu Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SubmenuOut"},"title":"Response Get Submenus Api V1 Menus  Menu Id  Submenus Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["submenus"],"summary":"Create Submenu","description":"Создать подменю","operationId":"create_submenu_api_v1_menus__menu_id__submenus_post","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","title":"Menu Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmenuIn"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmenuOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/menus/{menu_id}/submenus/{submenu_id}":{"get":{"tags":["submenus"],"summary":"Get Submenu","description":"Получчить подменю по id","operationId":"get_submenu_api_v1_menus__menu_id__submenus__submenu_id__get","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","title":"Menu Id"}},{"name":"submenu_id","in":"path","required":true,"schema":{"type":"string","title":"Submenu Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmenuOut"}}}},"404":{"description":"submenu not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["submenus"],"summary":"Update Submenu","description":"Обновить подменю","operationId":"update_submenu_api_v1_menus__menu_id__submenus__submenu_id__patch","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","title":"Menu Id"}},{"name":"submenu_id","in":"path","required":true,"schema":{"type":"string","title":"Submenu Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmenuIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmenuOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["submenus"],"summary":"Delete Submenu","description":"Удалить подменю","operationId":"delete_submenu_api_v1_menus__menu_id__submenus__submenu_id__delete","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","title":"Menu Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmenuOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/menus/{menu_id}/submenus/{submenu_id}/dishes":{"get":{"tags":["dishes"],"summary":"Get Dishes","description":"Получить список всех блюд","operationId":"get_dishes_api_v1_menus__menu_id__submenus__submenu_id__dishes_get","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","title":"Menu Id"}},{"name":"submenu_id","in":"path","required":true,"schema":{"type":"string","title":"Submenu Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DishOut"},"title":"Response Get Dishes Api V1 Menus  Menu Id  Submenus  Submenu Id  Dishes Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["dishes"],"summary":"Create Dish","description":"Создать блюдо","operationId":"create_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes_post","parameters":[{"name":"submenu_id","in":"path","required":true,"schema":{"title":"Submenu Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DishIn"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DishOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/menus/{menu_id}/submenus/{submenu_id}/dishes/{dish_id}":{"get":{"tags":["dishes"],"summary":"Get Dish","description":"Получить блюдо по id","operationId":"get_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__get","parameters":[{"name":"submenu_id","in":"path","required":true,"schema":{"title":"Submenu Id"}},{"name":"dish_id","in":"path","required":true,"schema":{"title":"Dish Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DishOut"}}}},"404":{"description":"dish not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["dishes"],"summary":"Update Dish","description":"Обновить блюдо","operationId":"update_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__patch","parameters":[{"name":"submenu_id","in":"path","required":true,"schema":{"title":"Submenu Id"}},{"name":"dish_id","in":"path","required":true,"schema":{"title":"Dish Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DishIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DishOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["dishes"],"summary":"Delete Dish","description":"Удалить блюдо","operationId":"delete_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__delete","parameters":[{"name":"dish_id","in":"path","required":true,"schema":{"title":"Dish Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DishOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"DishIn":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"price":{"type":"string","title":"Price"}},"type":"object","required":["title","description","price"],"title":"DishIn"},"DishOut":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"price":{"type":"string","title":"Price"},"id":{"type":"string","format":"uuid","title":"Id"}},"type":"object","required":["title","description","price","id"],"title":"DishOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MenuIn":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"}},"type":"object","required":["title","description"],"title":"MenuIn"},"MenuOut":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"id":{"type":"string","format":"uuid","title":"Id"},"submenus_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Submenus Count","default":0},"dishes_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Dishes Count","default":0}},"type":"object","required":["title","description","id"],"title":"MenuOut"},"SubmenuIn":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"}},"type":"object","required":["title","description"],"title":"SubmenuIn"},"SubmenuOut":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"id":{"type":"string","format":"uuid","title":"Id"},"dishes_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Dishes Count","default":0}},"type":"object","required":["title","description","id"],"title":"SubmenuOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}